Simplify access of repair_type_overrides, key by String
A lot of care existed to ensure each RepairType was populated in
repair_type_overrides (using ensureOverrides).

This could be simplified quite a bit by protecting access to
repair_type_overrides by enclosing it in a getOptions method that
ensures the provided RepairType has a populated entry.

Additionally, change the key type to String and use ConcurrentMap;
otherwise the change to make the enum names all caps does not work.