Check presence of a snapshot in a case-insensitive manner on macOS platform to prevent hardlinking failures
When hardlinks are created for a snapshot with the name "snapshot"
and then we take a snapshot with the name "Snapshot", macOS platform thinks
that this was already hardlinked because its hardlinking implementation
does not seem to be case-sensitive. The fix consists of checking,
in a case-insensitive manner, if there is already such snapshot,
but only on macOS platform.

The patch also does not create empty directories for a given snapshot
prematurely before checking its presence so there are not dangling empty
directories in case of a failure.

patch by Stefan Miklosovic; reviewed by Bernardo Botella, Caleb Rackliffe for CASSANDRA-20146