Update testAsyncPstmtInvalidation to be bound around evictions
Instead of preparing 10000 statements always, detect when we start
evicting, and then load as many more.

This allows the test to behave similarly to testPstmtInvalidation
as it will prevent unneeded additional loading of statmements which
adds execution time to the test.

This reduces the runtime of the test by ~3-5x depending on the
execution environment.